Refactor VM Op imports out of util/vsphere/vm
This patch removes the package imports from util/vsphere/vm that
reference the VM Op API or Kubernetes code. The package util/vsphere/vm
is meant to be pure vSphere code, and should never reference anything
related to Kubernetes.
