+++ pkg/providers/vsphere/virtualmachine/virtualmachine_suite_test.go
@@ -18,6 +18,8 @@ func vcSimTests() {
 	Describe("Publish", Label(testlabels.VCSim), publishTests)
 	Describe("Backup", Label(testlabels.VCSim), backupTests)
 	Describe("GuestInfo", Label(testlabels.VCSim), guestInfoTests)
+	Describe("CD-ROM", Label(testlabels.VCSim), cdromTests)
+	Describe("ExtraConfig", Label(testlabels.VCSim), extraConfigTests)
 }
 
 var suite = builder.NewTestSuite()
